requirements.txt
Flask==1.1.1
Flask_Bootstrap==3.3.7.1
oauth2client==4.1.3
gspread==3.6.0
PyYAML==5.3.1
